(hel-401) Utiliser composant react plutot qur view model pour les histogrammes horizontaux
