+++ src/frontend/ui/commun/Graphique/GraphiqueViewModel.tsx
@@ -112,132 +112,6 @@ export class GraphiqueViewModel {
     );
   }
 
-  protected afficheUnCarrousel(
-    chartColors: string[],
-    valeursDeGauche: number[],
-    valeursDeDroite: number[],
-    libellés: string[],
-    ratioLargeurSurHauteur: number,
-    entêtePremièreColonne: string,
-    entêtesDesAutresColonnes: string[],
-    annéesManquantes: number[] | string[],
-    nombreDAnnéeTotale: number = 3
-  ): ReactElement {
-    ChartJS.unregister();
-    const chartColorsGauche = valeursDeGauche.map((valeurDeGauche, index) => {
-      return valeurDeGauche <= 0 ? chartColors[index] : this.couleurDuFondHistogrammeDeDépassement;
-    });
-    const chartColorsDroite = valeursDeDroite.map((valeurDeDroite, index) => {
-      return valeurDeDroite >= 0 ? chartColors[index] : this.couleurDuFondHistogrammeDeDépassement;
-    });
-    const datalabelsColorsGauche = valeursDeGauche.map((valeurDeGauche) => {
-      return valeurDeGauche <= 0 ? this.couleurDeLaValeur : this.couleurDuFondHistogrammeDeDépassement;
-    });
-    const datalabelsColorsDroite = valeursDeDroite.map((valeurDeDroite) => {
-      return valeurDeDroite >= 0 ? this.couleurDeLaValeur : this.couleurDuFondHistogrammeDeDépassement;
-    });
-    const dataGauche: ChartData = {
-      datasets: [
-        {
-          backgroundColor: chartColorsGauche,
-          data: valeursDeGauche.map((valeurDeGauche) => Math.abs(valeurDeGauche)),
-          datalabels: {
-            font: { weight: "bold" },
-            formatter: (valeurDeGauche) => {
-              const valeurDeGaucheSignée = valeursDeGauche.includes(valeurDeGauche) ? valeurDeGauche : valeurDeGauche * -1;
-              return StringFormater.formateLeMontantEnEuros(valeurDeGaucheSignée);
-            },
-            labels: { title: { color: datalabelsColorsGauche } },
-          },
-          maxBarThickness: 60,
-          type: "bar",
-          yAxisID: "y",
-        },
-      ],
-      labels: libellés,
-    };
-    const dataDroite: ChartData = {
-      datasets: [
-        {
-          backgroundColor: chartColorsDroite,
-          data: valeursDeDroite,
-          datalabels: {
-            font: { weight: "bold" },
-            formatter: (valeurDeDroite) => StringFormater.formateLeMontantEnEuros(valeurDeDroite),
-            labels: { title: { color: datalabelsColorsDroite } },
-          },
-          maxBarThickness: 60,
-          type: "bar",
-          yAxisID: "y",
-        },
-      ],
-      labels: libellés,
-    };
-    const valeursTableau = [
-      [
-        StringFormater.formateLeMontantEnEuros(valeursDeGauche[0]),
-        StringFormater.formateLeMontantEnEuros(valeursDeGauche[1]),
-        StringFormater.formateLeMontantEnEuros(valeursDeGauche[2]),
-        StringFormater.formateLeMontantEnEuros(valeursDeGauche[3]),
-      ],
-      [
-        StringFormater.formateLeMontantEnEuros(valeursDeDroite[0]),
-        StringFormater.formateLeMontantEnEuros(valeursDeDroite[1]),
-        StringFormater.formateLeMontantEnEuros(valeursDeDroite[2]),
-        StringFormater.formateLeMontantEnEuros(valeursDeDroite[3]),
-      ],
-    ];
-
-    return (
-      <>
-        {annéesManquantes.length < nombreDAnnéeTotale && (
-          <div
-            style={{
-              display: "grid",
-              gridTemplateColumns: "repeat(2, 50%)",
-            }}
-          >
-            <div>
-              <Bar
-                // @ts-ignore
-                data={dataGauche}
-                options={this.optionsHistogrammeHorizontal(
-                  ratioLargeurSurHauteur,
-                  Math.max(...valeursDeGauche.map(Number), ...valeursDeDroite.map(Number)) * 1.1,
-                  ["400"],
-                  entêtesDesAutresColonnes[0]
-                )}
-                redraw={true}
-              />
-            </div>
-            <div>
-              <Bar
-                // @ts-ignore
-                data={dataDroite}
-                options={this.optionsHistogrammeHorizontal(
-                  ratioLargeurSurHauteur,
-                  Math.max(...valeursDeGauche.map(Number), ...valeursDeDroite.map(Number)),
-                  ["400"],
-                  entêtesDesAutresColonnes[1]
-                )}
-                redraw={true}
-              />
-            </div>
-          </div>
-        )}
-        {annéesManquantes.length > 0 && <MiseEnExergue>{`${this.wording.AUCUNE_DONNÉE_RENSEIGNÉE} ${annéesManquantes.join(", ")}`}</MiseEnExergue>}
-        <Transcription
-          disabled={annéesManquantes.length === nombreDAnnéeTotale}
-          entêteLibellé={entêtePremièreColonne}
-          identifiantUnique="compte-de-resultat"
-          identifiants={entêtesDesAutresColonnes}
-          libellés={libellés}
-          valeurs={valeursTableau}
-        />
-      </>
-    );
-  }
-
   protected afficheUnHistogrammeVertical(
     valeurs: number[],
     libellés: (number | string)[],
@@ -550,64 +424,6 @@ export class GraphiqueViewModel {
         y: {
           display: false,
           max: this.borneMaximaleDeLHistogrammeVertical,
-        },
-      },
-    };
-  }
-
-  private optionsHistogrammeHorizontal(
-    ratioLargeurSurHauteur: number,
-    valeurMaximale: number,
-    grosseursDePoliceDesLibellés: string[],
-    title: string = ""
-  ): ChartOptions<"bar"> {
-    return {
-      animation: false,
-      aspectRatio: ratioLargeurSurHauteur,
-      indexAxis: "y",
-      plugins: {
-        datalabels: {
-          align: "end",
-          anchor: "end",
-          font: {
-            family: "Marianne",
-            size: 14,
-          },
-          formatter: (value: string, _context: Context): string => parseFloat(value).toLocaleString("fr"),
-        },
-        legend: { display: false },
-        tooltip: { enabled: false },
-      },
-      scales: {
-        x: {
-          grid: {
-            display: false,
-            drawBorder: false,
-          },
-          max: 1.45 * (valeurMaximale > 0 ? valeurMaximale : 1),
-          min: 0,
-          position: "top",
-          ticks: { display: false },
-          title: {
-            align: "start",
-            color: this.couleurIdentifiant,
-            display: title === "" ? false : true,
-            font: { weight: "bold" },
-            text: title,
-          },
-        },
-        y: {
-          grid: {
-            drawBorder: false,
-            drawOnChartArea: false,
-            drawTicks: false,
-          },
-          ticks: {
-            color: this.couleurDelAbscisse,
-            // @ts-ignore
-            font: { weight: grosseursDePoliceDesLibellés },
-            padding: 8,
-          },
         },
       },
     };
